Update VERSION to reflect the latest patch release
Owing to the fact that we have started towards perfecting a workflow
that does not include merging release branches back into the default
branch, a discrepancy such as this stemmed as a side effect of the last
release.

In a conversation with @dgrisonnet, we saw that by adhering to this
process, backports will be much more cleaner and easier to maintain,
which is also how other sub-projects' release workflow is structured
currently.

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>